refactor: moved InPlaceDrop into in_place_drop.rs
